feat(ReviewFeed): Add styling, redux, & more
This commit adds basic styling for the Letterboxd RSS feed. Also, it
introduces the react-redux structure, with all components still
remaining functional. The refactor that places the /components/ folder
in the /src/ folder now. This allows grouping (for reading ease) of the
actions, reducer, store, & components folder within the /src/ folder.