feat(login): Add foundation of refresh token logic
Known Issue: Race condition in the refresh token & get user info that
causes the userinfo to not actually get updated. Only updates if
getUserInfo resolves last in the Layout component. Solution: Chain these
two actions together if needed, although that is likely not the
end solution for this, as the Auth token is refreshed while other
requests might be out. End solution is to place token refresh on a
middleware or something